requirements.txt
fastapi
uvicorn
pydantic
pynacl
requests
redis  
customtkinter
qrcode
pillow
psutil
PyJWT
passlib[bcrypt]
SQLAlchemy
python-dotenv
python-multipart